Create UniqueTokenImage component
