+++ src/components/asset-list/UniqueTokenCard.js
@@ -1,28 +1,17 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { compose, withHandlers, withProps, withState } from 'recompact';
 import styled from 'styled-components/primitives';
-import { fonts, padding, position, shadow } from '../../styles';
+import { position, shadow } from '../../styles';
 import { Centered } from '../layout';
-import { Monospace } from '../text';
 import { ShadowStack } from '../shadow-stack';
-import Shimmer from '../Shimmer';
+import UniqueTokenImage from './UniqueTokenImage';
 
 const UniqueTokenCardBorderRadius = 16;
 
 const Container = styled(Centered)`
-  ${padding(19, 10)}
   ${position.cover}
   border-radius: ${UniqueTokenCardBorderRadius};
   background-color: ${({ background }) => background};
-`;
-
-const FallbackText = styled(Monospace).attrs({
-  color: 'blueGreyLight',
-  size: 'smedium',
-})`
-  line-height: ${fonts.lineHeight.loose};
-  text-align: center;
 `;
 
 const InnerBorder = styled.View`
@@ -32,24 +21,7 @@ const InnerBorder = styled.View`
   border-width: 0.68;
 `;
 
-const UniqueTokenImage = styled.Image`
-  ${position.size('100%')}
-`;
-
-const UniqueTokenCard = ({
-  item: {
-    background,
-    imageUrl,
-  },
-  error,
-  isLoading,
-  name,
-  onError,
-  onLoad,
-  onLoadStart,
-  size,
-  ...props
-}) => (
+const UniqueTokenCard = ({ item: { background, imagePreviewUrl, ...item }, size, ...props }) => (
   <ShadowStack
     {...props}
     {...position.sizeAsObject(size)}
@@ -60,49 +32,22 @@ const UniqueTokenCard = ({
     ]}
   >
     <Container background={background}>
-      {(imageUrl && !error) ? (
-        <UniqueTokenImage
-          onError={onError}
-          onLoad={onLoad}
-          onLoadStart={onLoadStart}
-          resizeMode="contain"
-          source={{ uri: imageUrl }}
-        />
-      ) : (
-        <FallbackText>{name}</FallbackText>
-      )}
-      {isLoading && <Shimmer {...position.sizeAsObject(size)} />}
+      <UniqueTokenImage
+        imageUrl={imagePreviewUrl}
+        item={item}
+        size={size}
+      />
       <InnerBorder />
     </Container>
   </ShadowStack>
 );
 
 UniqueTokenCard.propTypes = {
-  error: PropTypes.object,
-  isLoading: PropTypes.bool,
   item: PropTypes.shape({
     background: PropTypes.string,
-    imageUrl: PropTypes.string,
+    imagePreviewUrl: PropTypes.string,
   }),
-  name: PropTypes.string.isRequired,
-  onError: PropTypes.func,
-  onLoad: PropTypes.func,
-  onLoadStart: PropTypes.func,
   size: PropTypes.number,
 };
 
-export default compose(
-  withState('error', 'handleErrorState', null),
-  withState('isLoading', 'handleLoadingState', false),
-  withHandlers({
-    onError: ({ handleErrorState, handleLoadingState }) => (error) => {
-      handleErrorState(error);
-      handleLoadingState(false);
-    },
-    onLoad: ({ handleLoadingState }) => () => handleLoadingState(false),
-    onLoadStart: ({ handleLoadingState }) => () => handleLoadingState(true),
-  }),
-  withProps(({ item: { contractName, id, name } }) => ({
-    name: name || `${contractName} #${id}`,
-  })),
-)(UniqueTokenCard);
+export default UniqueTokenCard;
