+++ src/components/asset-list/index.js
@@ -3,4 +3,5 @@ export { default as AssetListFooter } from './AssetListFooter';
 export { default as AssetListHeader } from './AssetListHeader';
 export { default as AssetListItem } from './AssetListItem';
 export { default as UniqueTokenCard } from './UniqueTokenCard';
+export { default as UniqueTokenImage } from './UniqueTokenImage';
 export { default as UniqueTokenRow } from './UniqueTokenRow';
